+++ src/snyk/common/constants/languageServer.ts
@@ -1,0 +1,8 @@
+// Language Server constants
+export const SNYK_LANGUAGE_SERVER_NAME = 'Snyk Language Server';
+
+// protocol methods
+export const DID_CHANGE_CONFIGURATION_METHOD = 'workspace/didChangeConfiguration';
+
+// custom methods
+export const SNYK_HAS_AUTHENTICATED = '$/snyk.hasAuthenticated';